Improve delete method -- return false if value to be deleted does not already exist

+++ RevSpaceWebService/src/main/java/com/revature/revspace/services/CrudService.java
@@ -84,6 +84,10 @@ public interface CrudService<T, KEY, REPO extends CrudRepository<T, KEY>>
 	{
 		try
 		{
+			if (!this.getRepo().findById(id).isPresent())
+			{
+				return false;
+			}
 			this.getRepo().deleteById(id);
 			return true;
 		}
